refactor: upgrade test coverage
